% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbc_lm.R
\name{cbc_lm}
\alias{cbc_lm}
\title{Case-by-Case Linear Regresssion}
\usage{
cbc_lm(
  data,
  formula,
  .case,
  n_bootstrap = 4000,
  lm_options = list(),
  boot_options = list(),
  boot.ci_options = list(),
  na.rm = FALSE
)
}
\arguments{
\item{data}{A data frame containing the variables in the model}

\item{formula}{An object of class formula or a string coercible to an object
of class formula}

\item{.case}{A quoted variable name used to subset data into cases}

\item{n_bootstrap}{The number of bootstrap replicates for standard errors and
confidence intervals of mean coefficients}

\item{lm_options}{Pass additional arguments to lm}

\item{boot_options}{Pass additional arguments to boot}

\item{boot.ci_options}{Pass additional arguments to boot.ci}

\item{na.rm}{Pass na.rm to: the mean function used to obtain mean_coef and bm_coef;
the sd function used to obtain se_coef; the mean function used in the statistic
parameter of boot}
}
\value{
An object of class cbc_lm
}
\description{
Case-by-Case Linear Regresssion
}
\examples{
  df <- data.frame(ids = rep(1:5, 5),
                   vals = stats::rnorm(25),
                   outs = stats::rnorm(25, 10, 25))

 cbc_lm(data = df, formula = outs ~ vals, .case = "ids")
}
