% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbc_lm.R
\name{print.cbc_lm}
\alias{print.cbc_lm}
\title{Print Method for 'cbc_lm' Objects}
\usage{
\method{print}{cbc_lm}(x, digits = max(3L, getOption("digits") - 3L), boot = FALSE, ...)
}
\arguments{
\item{x}{A 'cbc_lm' object.}

\item{digits}{The number of significant digits to use when printing.}

\item{boot}{Logical indicating whether or not to print the bootstrap mean
coefficients.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
An invisible 'cbc_lm' object.
}
\description{
Print method for 'cbc_lm' objects. Shows the call used to create the model,
the mean coefficients, (optionally) the bootstrap mean coefficients,
and the coefficients for each model.
}
\seealso{
\code{\link{summary.cbc_lm}}, \code{\link{plot.cbc_lm}}
}
