% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbc_lm.R
\name{print.summary.cbc_lm}
\alias{print.summary.cbc_lm}
\title{Print Method for 'summary.cbc_lm' Objects}
\usage{
\method{print}{summary.cbc_lm}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{A 'summary.cbc_lm' object.}

\item{digits}{The number of significant digits to use when printing.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
An invisible 'summary.cbc_lm' object.
}
\description{
Print method for 'summary.cbc_lm' objects. Prints the call used to create the
models, the mean coefficients, (optionally) the bootstrap mean coefficients,
bootstrap standard errors, bootstrap confidence intervals,
and the tidy and glance summaries for each model.
}
\seealso{
\code{\link{print.cbc_lm}}, \code{\link{summary.cbc_lm}}
}
