% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{robins}
\alias{robins}
\title{Ratio of robin males to females in Walker Creek and Knobs Flat, Eglinton Valley}
\format{
\subsection{\code{robins}}{

A data frame with 2 observations and 6 variables:
\describe{
\item{site}{Site name}
\item{aug_05, aug_06, aug_07, aug_08, aug_09}{ratio of male to female robins}
}
}
}
\source{
\url{https://www.doc.govt.nz/documents/science-and-technical/inventory-monitoring/im-toolbox-birds-incomplete-five-min-counts.pdf}
}
\usage{
robins
}
\description{
Data from Table 1 of "Birds: incomplete counts—five-minute bird counts Version 1.0"
}
\keyword{datasets}
